Moved notifier to general components
