+++ lib/assets/javascripts/cartodb3/components/background-importer/background-geocoding-item-view.js
@@ -3,7 +3,7 @@ var _ = require('underscore');
 var Utils = require('cdb.Utils');
 var GeocodingResultDetailsView = require('../geocoding/geocoding-result-details-view');
 var template = require('./background-geocoding-item.tpl');
-var Notifier = require('../../editor/components/notifier/notifier');
+var Notifier = require('../../components/notifier/notifier');
 
 /**
  *  Geocoding item within background polling
