+++ lib/assets/javascripts/cartodb3/components/background-importer/background-import-item-view.js
@@ -1,5 +1,5 @@
 var CoreView = require('backbone/core-view');
-var Notifier = require('../../editor/components/notifier/notifier');
+var Notifier = require('../../components/notifier/notifier');
 var UploadConfig = require('../../config/upload-config');
 var ErrorDetailsView = require('./error-details-view');
 var WarningsDetailsView = require('./warnings-details-view');
