+++ lib/assets/javascripts/cartodb3/components/background-importer/background-import-limit-view.js
@@ -1,6 +1,6 @@
 var CoreView = require('backbone/core-view');
 var template = require('./background-import-limit.tpl');
-var Notifier = require('../../editor/components/notifier/notifier');
+var Notifier = require('../../components/notifier/notifier');
 
 /**
  *  Import limit message within background importer
