+++ lib/assets/javascripts/cartodb3/components/notifier/notifier-action-view.js
