+++ lib/assets/javascripts/cartodb3/components/notifier/notifier-collection.js
