+++ lib/assets/javascripts/cartodb3/components/notifier/notifier-item-view.js
