+++ lib/assets/javascripts/cartodb3/components/table/operations/table-rename-column.js
@@ -1,4 +1,4 @@
-var Notifier = require('../../../editor/components/notifier/notifier');
+var Notifier = require('../../../components/notifier/notifier');
 var errorParser = require('../../../helpers/error-parser');
 
 module.exports = function (opts) {
