+++ lib/assets/javascripts/cartodb3/components/table/paginator/table-paginator-view.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var CoreView = require('backbone/core-view');
 var template = require('./table-paginator.tpl');
 var templateLoader = require('./table-paginator-loader.tpl');
-var Notifier = require('../../../editor/components/notifier/notifier');
+var Notifier = require('../../../components/notifier/notifier');
 var errorParser = require('../../../helpers/error-parser');
 
 module.exports = CoreView.extend({
