+++ lib/assets/javascripts/cartodb3/editor.js
@@ -34,7 +34,7 @@ var ImporterManager = require('./components/background-importer/background-impor
 var BackgroundPollingModel = require('./data/editor-background-polling-model');
 var StyleManager = require('./editor/style/style-manager');
 var DeepInsightsIntegrations = require('./deep-insights-integrations');
-var Notifier = require('./editor/components/notifier/notifier');
+var Notifier = require('./components/notifier/notifier');
 
 // JSON data passed from entry point (editor/visualizations/show.html):
 var vizJSON = window.vizJSON;
