+++ lib/assets/javascripts/cartodb3/editor/editor-map-view.js
@@ -6,7 +6,7 @@ var StackLayoutView = require('../components/stack-layout/stack-layout-view');
 var BasemapContentView = require('./layers/basemap-content-view');
 var LayerContentView = require('./layers/layer-content-view');
 var WidgetsFormContentView = require('./widgets/widgets-form/widgets-form-content-view');
-var Notifier = require('./components/notifier/notifier');
+var Notifier = require('../components/notifier/notifier');
 
 module.exports = CoreView.extend({
   className: 'MapEditor Editor-panel',
