+++ lib/assets/javascripts/cartodb3/editor/layers/operations/rename-layer.js
@@ -1,4 +1,4 @@
-var Notifier = require('../../../editor/components/notifier/notifier');
+var Notifier = require('../../../components/notifier/notifier');
 var errorParser = require('../../../helpers/error-parser');
 
 module.exports = function (opts) {
