+++ lib/assets/test/spec/cartodb3/components/notifier/notifier-collection.spec.js
@@ -1,6 +1,6 @@
-var NotifierCollection = require('../../../../../../javascripts/cartodb3/editor/components/notifier/notifier-collection.js');
+var NotifierCollection = require('../../../../../javascripts/cartodb3/components/notifier/notifier-collection.js');
 
-describe('editor/components/notifier/notifier-collection', function () {
+describe('components/notifier/notifier-collection', function () {
   beforeEach(function () {
     this.collection = new NotifierCollection();
     this.collection.add({
