+++ lib/assets/test/spec/cartodb3/components/notifier/notifier-view.spec.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
-var Notifier = require('../../../../../../javascripts/cartodb3/editor/components/notifier/notifier.js');
+var Notifier = require('../../../../../javascripts/cartodb3/components/notifier/notifier.js');
 
-describe('editor/components/notifier/notifier-view', function () {
+describe('components/notifier/notifier-view', function () {
   var view;
   var editorModel;
