+++ lib/assets/test/spec/cartodb3/components/notifier/notifier.spec.js
@@ -1,7 +1,7 @@
-var Notifier = require('../../../../../../javascripts/cartodb3/editor/components/notifier/notifier.js');
+var Notifier = require('../../../../../javascripts/cartodb3/components/notifier/notifier.js');
 var Backbone = require('backbone');
 
-describe('editor/components/notifier/notifier', function () {
+describe('components/notifier/notifier', function () {
   var model;
   var editorModel;
