+++ lib/assets/test/spec/cartodb3/editor/editor-map-view.spec.js
@@ -9,7 +9,7 @@ var ModalsService = require('../../../../javascripts/cartodb3/components/modals/
 var EditorMapView = require('../../../../javascripts/cartodb3/editor/editor-map-view');
 var EditorModel = require('../../../../javascripts/cartodb3/data/editor-model');
 var VisDefinitionModel = require('../../../../javascripts/cartodb3/data/vis-definition-model');
-var Notifier = require('../../../../javascripts/cartodb3/editor/components/notifier/notifier');
+var Notifier = require('../../../../javascripts/cartodb3/components/notifier/notifier');
 
 describe('editor/editor-map-view', function () {
   beforeEach(function () {
